refactor(rbac): move platform role SQL into a repository
PlatformRoleService no longer depends on TypeORM. The SQL lives in
PlatformRoleRepository, which uses TransactionService.getManager() like
the other repositories, so it participates in the ambient transaction.
