+++ apps/api/src/domains/rbac/rbac.module.ts
@@ -3,6 +3,7 @@ import { ConfigModule } from "@nestjs/config"
 import { TypeOrmModule } from "@nestjs/typeorm"
 import { CheckPermissionGuard } from "./check-permission.guard"
 import { PermissionService } from "./permission.service"
+import { PlatformRoleRepository } from "./platform-role.repository"
 import { PlatformRoleService } from "./platform-role.service"
 import { RbacService } from "./rbac.service"
 import { Role } from "./role.entity"
@@ -10,7 +11,13 @@ import { RolePermission } from "./role-permission.entity"
 
 @Module({
   imports: [ConfigModule, TypeOrmModule.forFeature([Role, RolePermission])],
-  providers: [RbacService, PermissionService, CheckPermissionGuard, PlatformRoleService],
+  providers: [
+    RbacService,
+    PermissionService,
+    CheckPermissionGuard,
+    PlatformRoleRepository,
+    PlatformRoleService,
+  ],
   exports: [RbacService, PermissionService, CheckPermissionGuard, PlatformRoleService],
 })
 export class RbacModule {}
